+++ src/ActivityLogger.php
@@ -5,9 +5,9 @@ namespace Spatie\Activitylog;
 use Illuminate\Auth\AuthManager;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Traits\Macroable;
-use Spatie\Activitylog\Contracts\Activity;
 use Illuminate\Contracts\Config\Repository;
 use Spatie\Activitylog\Exceptions\CouldNotLogActivity;
+use Spatie\Activitylog\Contracts\Activity as ActivityContract;
 
 class ActivityLogger
 {
@@ -16,16 +16,7 @@ class ActivityLogger
     /** @var \Illuminate\Auth\AuthManager */
     protected $auth;
 
-    protected $logName = '';
-
-    /** @var \Illuminate\Database\Eloquent\Model */
-    protected $performedOn;
-
-    /** @var \Illuminate\Database\Eloquent\Model */
-    protected $causedBy;
-
-    /** @var \Illuminate\Support\Collection */
-    protected $properties;
+    protected $defaultLogName = '';
 
     /** @var string */
     protected $authDriver;
@@ -33,23 +24,16 @@ class ActivityLogger
     /** @var \Spatie\Activitylog\ActivityLogStatus */
     protected $logStatus;
 
+    /** @var \Spatie\Activitylog\Contracts\Activity */
+    protected $activity;
+
     public function __construct(AuthManager $auth, Repository $config, ActivityLogStatus $logStatus)
     {
         $this->auth = $auth;
 
-        $this->properties = collect();
-
         $this->authDriver = $config['activitylog']['default_auth_driver'] ?? $auth->getDefaultDriver();
 
-        if (starts_with(app()->version(), '5.1')) {
-            $this->causedBy = $auth->driver($this->authDriver)->user();
-        } else {
-            $this->causedBy = $auth->guard($this->authDriver)->user();
-        }
-
-        $this->logName = $config['activitylog']['default_log_name'];
-
-        $this->logEnabled = $config['activitylog']['enabled'] ?? true;
+        $this->defaultLogName = $config['activitylog']['default_log_name'];
 
         $this->logStatus = $logStatus;
     }
@@ -63,7 +47,7 @@ class ActivityLogger
 
     public function performedOn(Model $model)
     {
-        $this->performedOn = $model;
+        $this->getActivity()->subject()->associate($model);
 
         return $this;
     }
@@ -81,7 +65,7 @@ class ActivityLogger
 
         $model = $this->normalizeCauser($modelOrId);
 
-        $this->causedBy = $model;
+        $this->getActivity()->causer()->associate($model);
 
         return $this;
     }
@@ -93,21 +77,21 @@ class ActivityLogger
 
     public function withProperties($properties)
     {
-        $this->properties = collect($properties);
+        $this->getActivity()->properties = collect($properties);
 
         return $this;
     }
 
     public function withProperty(string $key, $value)
     {
-        $this->properties->put($key, $value);
+        $this->getActivity()->properties = $this->getActivity()->properties->put($key, $value);
 
         return $this;
     }
 
     public function useLog(string $logName)
     {
-        $this->logName = $logName;
+        $this->getActivity()->log_name = $logName;
 
         return $this;
     }
@@ -115,6 +99,13 @@ class ActivityLogger
     public function inLog(string $logName)
     {
         return $this->useLog($logName);
+    }
+
+    public function tap(callable $callback, string $eventName = null)
+    {
+        call_user_func($callback, $this->getActivity(), $eventName);
+
+        return $this;
     }
 
     public function enableLogging()
@@ -137,23 +128,13 @@ class ActivityLogger
             return;
         }
 
-        $activity = ActivitylogServiceProvider::getActivityModelInstance();
-
-        if ($this->performedOn) {
-            $activity->subject()->associate($this->performedOn);
-        }
-
-        if ($this->causedBy) {
-            $activity->causer()->associate($this->causedBy);
-        }
-
-        $activity->properties = $this->properties;
+        $activity = $this->activity;
 
         $activity->description = $this->replacePlaceholders($description, $activity);
 
-        $activity->log_name = $this->logName;
-
         $activity->save();
+
+        $this->activity = null;
 
         return $activity;
     }
@@ -164,11 +145,7 @@ class ActivityLogger
             return $modelOrId;
         }
 
-        if (starts_with(app()->version(), '5.1')) {
-            $model = $this->auth->driver($this->authDriver)->getProvider()->retrieveById($modelOrId);
-        } else {
-            $model = $this->auth->guard($this->authDriver)->getProvider()->retrieveById($modelOrId);
-        }
+        $model = $this->auth->guard($this->authDriver)->getProvider()->retrieveById($modelOrId);
 
         if ($model) {
             return $model;
@@ -177,7 +154,7 @@ class ActivityLogger
         throw CouldNotLogActivity::couldNotDetermineUser($modelOrId);
     }
 
-    protected function replacePlaceholders(string $description, Activity $activity): string
+    protected function replacePlaceholders(string $description, ActivityContract $activity): string
     {
         return preg_replace_callback('/:[a-z0-9._-]+/i', function ($match) use ($activity) {
             $match = $match[0];
@@ -200,5 +177,18 @@ class ActivityLogger
 
             return array_get($attributeValue, $propertyName, $match);
         }, $description);
+    }
+
+    protected function getActivity(): ActivityContract
+    {
+        if (! $this->activity instanceof ActivityContract) {
+            $this->activity = ActivitylogServiceProvider::getActivityModelInstance();
+            $this
+                ->useLog($this->defaultLogName)
+                ->withProperties([])
+                ->causedBy($this->auth->guard($this->authDriver)->user());
+        }
+
+        return $this->activity;
     }
 }